+++ leechbot/commands/__init__.py
@@ -4,6 +4,7 @@
 from . import admin
 from . import downloads
 from . import options
+from . import rss
 from . import settings
 from . import start_help
 from . import status